backend/src/requirements.txt
alembic==1.11.1
anyio==3.7.0
bcrypt==4.0.1
cachetools==4.2.4
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.95.2
google-api-core==1.34.0
google-auth==1.35.0
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-cloud-texttospeech==2.14.1
google-cloud-translate==3.11.1
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpcio==1.54.2
grpcio-status==1.48.2
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
passlib==1.7.4
password-strength==0.0.3.post2
proto-plus==1.22.2
protobuf==3.20.3
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.8
python-dotenv==1.0.0
python-jose==3.3.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
typing_extensions==4.6.2
urllib3==1.26.16
uvicorn==0.22.0
